Improve functionality significantly
I'm still getting the hang of concurrency in Python in general.

Asyncio in particular seems to have some quirks.

Things seem to be working fairly well, but I think I've set
up the async context in such a way as to drop the BLE
connection after not much time. Will dig deeper.